Fix issue with quick launch not working
Previously, the `use-http`/`useFetch` hook was attached to a
`QuickLaunchButton`.  That component was being mounted and unmounted as
the `LaunchDropdown` was re-rendered.  I suspect that the unmounting was
caused by a different `toggle` function being passed to `ButtonDropdown`
on each render.

Usually, this would not make any practical difference, however the
`useFetch` hook is cancelled when its component is unmounted.  That
cancellation was erroneously picked up as a failed attempt to launch the
desktop.

The solution I've implemented has been to fold the `QuickLaunchButton`
into the new `LaunchDropdown`.  This has allowed hoisting the `useFetch`
hook to a component that is not unmounted.

An alternative, might have been to look at preventing the unmounting,
but hoisting the `useFetch` hook seemed like a more robust solution.
